backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
python-dotenv
supabase-py>=1.0,<2.0
httpx>=0.20.0,<1.0.0
